pkg/charts/loaded: Disable the TestLoadDirWithDevNull test
This test (and potentially any of the other symlink-related unit tests) is problematic with recent changes made to Cachito, as the `testdata/frobnitz_with_dev_null/null` links to /dev/null, which is outside of the local helm fork filesystem.

We don't rely on these unit tests ourselves, so this should be fine to disable for now.